requirements.txt
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
Werkzeug==1.0.1
torch
torchvision
gevent